ship_o_cereal/requirements.txt
appdirs==1.4.4
appnope==0.1.0
asgiref==3.3.4
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.9.11
chardet==4.0.0
decorator==4.4.1
defusedxml==0.6.0
distlib==0.3.1
Django==3.2.1
django-countries==7.2.1
django-credit-cards==0.4.1
django-crispy-forms==1.11.2
django-environ==0.4.5
entrypoints==0.3
filelock==3.0.12
idna==2.10
importlib-metadata==0.23
importlib-resources==3.0.0
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
MarkupSafe==1.1.1
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==2.0.10
psycopg2-binary==2.8.6
ptyprocess==0.6.0
Pygments==2.4.2
pyrsistent==0.15.6
python-dateutil==2.8.1
pytz==2021.1
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.25.1
Send2Trash==1.5.0
six==1.13.0
sqlparse==0.4.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
typing-extensions==3.10.0.0
urllib3==1.26.4
virtualenv==20.0.31
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==0.6.0
